Changed floatingBufferCount to double for sample readout, fixes stepped speed dial and premature endings for longer samples

+++ Assets/Scripts/CoreClasses/clipPlayer.cs
@@ -41,7 +41,7 @@ public class clipPlayer : signalGenerator {
   public float[] clipSamples;
   public int clipChannels = 2;
   public int[] sampleBounds = new int[] { 0, 0 };
-  public float floatingBufferCount = 0;
+  public double floatingBufferCount = 0;
   public int bufferCount = 0;
   public Vector2 trackBounds = new Vector2(0, 1);
